server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "pydantic-settings>=2.7.1",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.22",
    "tavily-python>=0.5.0",
    "trafilatura>=2.0.0",
    "sentence-transformers>=3.3.1",
    "numpy>=2.2.1",
    "google-generativeai>=0.8.3",
]
